+++ src/table/table-toolbar/index.js
@@ -2,6 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 import Toolbar from '@material-ui/core/Toolbar'
+import Typography from '@material-ui/core/Typography'
 import { makeStyles } from '@material-ui/core/styles'
 
 import Download from './download'
@@ -9,13 +10,17 @@ import Toggle from './toggle'
 import DefaultFilter from '../filters/default-filter'
 
 
-const useStyles = makeStyles(() => ({
+const useStyles = makeStyles((theme) => ({
   grow: {
     flexGrow: 1,
+  },
+  title: {
+    marginRight: theme.spacing(2),
   },
 }))
 
 const TableToolbar = ({
+  title,
   // Search
   preGlobalFilteredRows,
   globalFilter,
@@ -32,6 +37,11 @@ const TableToolbar = ({
 
   return (
     <Toolbar>
+      {title && (
+        <Typography className={classes.title} variant='h6' id='tableTitle'>
+          {title}
+        </Typography>
+      )}
       {allColumns.some((c) => !c.disableGlobalFilter) && (
         <DefaultFilter
           preFilteredRows={preGlobalFilteredRows}
@@ -58,10 +68,15 @@ const TableToolbar = ({
 }
 
 TableToolbar.propTypes = {
+  title: PropTypes.string,
   preGlobalFilteredRows: PropTypes.array.isRequired,
   globalFilter: PropTypes.string.isRequired,
   setGlobalFilter: PropTypes.func.isRequired,
   ...Download.propTypes,
   ...Toggle.propTypes,
 }
+TableToolbar.defaultProps = {
+  title: '',
+}
+
 export default TableToolbar